requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
pymongo==4.6.0
dnspython==2.4.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator>=2.0.0
